# [IMP] product: improve name_get performance