@rco-odoo 2022-01-26
@ryv-odoo sorry for not transcribing the discussion I had with @Aurelienvd yesterday.
I suggested another fix that basically avoids "partitioning" the prefetch set of ptavs, and that's a much simpler fix.
Aurélien is currently testing this alternative.