@Aurelienvd 2022-01-26
@rco-odoo @ryv-odoo 

So I tested the suggested fix. For ryv, the fix is simply changing `_get_combination_name` code to
```python
ptavs = self._without_no_variant_attributes().with_prefetch(self._prefetch_ids)
ptavs = ptavs._filter_single_value_lines().with_prefetch(self._prefetch_ids)
return ", ".join(ptav.name for ptav in ptavs)
```

I benchmarked everything with the customer DB (30K products).

| Products | Before PR | After PR (map) | After PR (prefetch) |
|:-----------:|:--------------:|:--------------------:|:-------------------------:|
| 25 | 0.045s | 0.029s | 0.029s |
| 100 | 0.132s | 0.046s | 0.044s |
| 500 | 0.493s | 0.1s | 0.128s |
| 2000 | 1.72s | 0.329s | 0.414s |
| 10000 | 5.74s | 1.58s | 2.068s |
| 30000 | 24.63s | 5.59s | 10.56s |

I tried to increase the size of `product.product_template_attribute_value_ids` but the changes
in exec. time weren't significant enough to report.

On average, excluding the 1 product case, the map fix gives a 3.6 speed-up while the prefetch fix
gives a 2.9 speed-up. The difference becomes noticeable around 2.5k products.