@ryv-odoo 2022-01-27
Hello @Aurelienvd ,

This suggested change of `_get_combination_name` seems the best option (even if the gain is smaller with a lot of record).
But, I don't understand the second prefetch (`ptavs._filter_single_value_lines().with_prefetch(self._prefetch_ids)`), because at this point, you only need the `name` which is already prefetch by the `_without_no_variant_attributes` or `_filter_single_value_lines` anyway. Or I miss something ?