@rco-odoo 2022-01-27
> But, I don't understand the second prefetch (`ptavs._filter_single_value_lines().with_prefetch(self._prefetch_ids)`), because at this point, you only need the `name` which is already prefetch by the `_without_no_variant_attributes` or `_filter_single_value_lines` anyway. Or I miss something ?

It's only a safety net, just in case we need to fetch fields, they will be fetched on a larger recordset.