@rco-odoo 2022-01-27
@Aurelienvd I also prefer the `with_prefetch()` patch in method `_get_combination_name()`, because it is simpler and much less invasive.  Please correct me if I am wrong, but I don't think we actually _need_ to invoke `name_get()` on several thousands of records, so the performance gain looks good enough for most cases.