@Aurelienvd 2022-01-27
@rco-odoo I also do think that the prefetch patch is better and does not mess with the context.
I put the benchmark here for the sake of completion.

Regarding the name_get on big recordsets, I think it can happen when loading the Forecasted Inventory Report for instance.
As it is a GridView with `row_fields = ["product_id", "state"]`, it could call name_get on a large recordset to display
those rows. But given that it is a time window-based view, I think that calling name_get on more than
10k products should still be pretty rare.